requirements.txt
fastapi
passlib
pydantic
python-jose  

